requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
requests
audio-recorder-streamlit
SpeechRecognition